Module atmega1280_hal::atmega1280

ATmega1280 Peripheral access API for ATMEGA1280 microcontrollers (generated using svd2rust v0.17.0 (5b392b4f3451 2020-08-15))

You can find an overview of the generated API here.

API features to be included in the next svd2rust release can be generated by cloning the svd2rust repository, checking out the above commit, and running cargo doc --open.

Modules

ac
Analog Comparator

adc
Analog-to-Digital Converter

boot_load
Bootloader

cpu
CPU Registers

eeprom
EEPROM

exint
External Interrupts

fuse
Fuses

jtag
JTAG Interface

lockbit
Lockbits

porta
I/O Port

portb
I/O Port

portc
I/O Port

portd
I/O Port

porte
I/O Port

portf
I/O Port

portg
I/O Port

porth
I/O Port

portj
I/O Port

portk
I/O Port

portl
I/O Port

spi
Serial Peripheral Interface

tc0
Timer/Counter, 8-bit

tc1
Timer/Counter, 16-bit

tc2
Timer/Counter, 8-bit Async

tc3
Timer/Counter, 16-bit

tc4
Timer/Counter, 16-bit

tc5
Timer/Counter, 16-bit

twi
Two Wire Serial Interface

usart0
USART

usart1
USART

usart2
USART

usart3
USART

wdt
Watchdog Timer

Structs

AC
Analog Comparator

ADC
Analog-to-Digital Converter

BOOT_LOAD
Bootloader

CPU
CPU Registers

EEPROM
EEPROM

EXINT
External Interrupts

FUSE
Fuses

JTAG
JTAG Interface

LOCKBIT
Lockbits

PORTA
I/O Port

PORTB
I/O Port

PORTC
I/O Port

PORTD
I/O Port

PORTE
I/O Port

PORTF
I/O Port

PORTG
I/O Port

PORTH
I/O Port

PORTJ
I/O Port

PORTK
I/O Port

PORTL
I/O Port

Peripherals
All the peripherals

SPI
Serial Peripheral Interface

TC0
Timer/Counter, 8-bit

TC1
Timer/Counter, 16-bit

TC2
Timer/Counter, 8-bit Async

TC3
Timer/Counter, 16-bit

TC4
Timer/Counter, 16-bit

TC5
Timer/Counter, 16-bit

TWI
Two Wire Serial Interface

USART0
USART

USART1
USART

USART2
USART

USART3
USART

WDT
Watchdog Timer

Enums

Interrupt
Enumeration of all the interrupts